refactor: move useAllowedUsers hook to src/hooks
